Major refactoring; particularly with factory design patterns for implementing native, GSL or BLAS/LAPACK functionality.
This is also reflected in the CMakeLists files to reflect this change.

+++ src/include/nrr_func.h
@@ -4,6 +4,6 @@ Num. Rec. functions
 
 real brent(real , real , real , real(*)(real), real , real *);
 real f1dim(real );
-void linmin(real *, real *, int , real *, real (*)() );
+void linmin(real *, real *, size_t , real *, real (*)() );
 void mnbrak(real *, real *, real *, real *, real *, real *, real (*)(real) );
 real ran1(long *);
